Stabilize most serde representations
