+++ src/offset_date_time.rs
@@ -22,27 +22,13 @@ use std::time::SystemTime;
 // [`PrimitiveDateTime`] coupled with a [`UtcOffset`]. This offset is added to
 // the date, time, or datetime as necessary for presentation or returning from a
 // function.
-#[cfg_attr(serde, derive(serde::Serialize))]
-#[cfg_attr(serde, serde(into = "crate::serde::PrimitiveDateTime"))]
+#[cfg_attr(serde, derive(serde::Serialize, serde::Deserialize))]
 #[derive(Debug, Clone, Copy, Eq)]
 pub struct OffsetDateTime {
     /// The `PrimitiveDateTime`, which is _always_ UTC.
     pub(crate) utc_datetime: PrimitiveDateTime,
     /// The `UtcOffset`, which will be added to the `PrimitiveDateTime` as necessary.
     pub(crate) offset: UtcOffset,
-}
-
-#[cfg(serde)]
-impl<'a> serde::Deserialize<'a> for OffsetDateTime {
-    #[inline(always)]
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'a>,
-    {
-        crate::serde::PrimitiveDateTime::deserialize(deserializer)?
-            .try_into()
-            .map_err(serde::de::Error::custom)
-    }
 }
 
 impl OffsetDateTime {
